+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+- Add multi-chart tracker layer
+- Add dashed line support
+- Migrate to Swift 2.3
+- Migrate to Swift 3.0
+- Fix division by 0 in axis layer
 
 ## [0.4] - 2016.05.29
 - Fix tracker showing NaN on constant coordinate